feat: increase halo2 circuit partitions

+++ filecoin-proofs/tests/mod.rs
@@ -33,7 +33,7 @@ fn test_verify_seal_fr32_validation() {
                         .read()
                         .expect("POREP_PARTITIONS poisoned")
                         .get(&SECTOR_SIZE_2_KIB)
-                        .expect("unknown sector size"),
+                        .expect("unknown sector size") as usize,
                 ),
                 porep_id: arbitrary_porep_id,
                 api_version: ApiVersion::V1_1_0,
@@ -72,7 +72,7 @@ fn test_verify_seal_fr32_validation() {
                         .read()
                         .expect("POREP_PARTITIONS poisoned")
                         .get(&SECTOR_SIZE_2_KIB)
-                        .expect("unknown sector size"),
+                        .expect("unknown sector size") as usize,
                 ),
                 porep_id: arbitrary_porep_id,
                 api_version: ApiVersion::V1_1_0,
@@ -115,7 +115,7 @@ fn test_verify_seal_fr32_validation() {
                         .read()
                         .expect("POREP_PARTITIONS poisoned")
                         .get(&SECTOR_SIZE_2_KIB)
-                        .expect("unknown sector size"),
+                        .expect("unknown sector size") as usize,
                 ),
                 porep_id: arbitrary_porep_id,
                 api_version: ApiVersion::V1_1_0,
